added plotting.py and use imputation in helpers
