+++ src/models.py
@@ -1,15 +1,69 @@
 import numpy as np
+import matplotlib.pyplot as plt
 from sklearn.linear_model import Ridge
 from sklearn.linear_model import GammaRegressor
-from sklearn.preprocessing import PolynomialFeatures
+from sklearn.preprocessing import PolynomialFeatures, StandardScaler
 from sklearn.pipeline import make_pipeline
 from sklearn.ensemble import RandomForestRegressor
 from sklearn.neighbors import KNeighborsRegressor
-from sklearn.neighbors.regression import check_array, _get_weights
+from sklearn.model_selection import train_test_split
+from sklearn.metrics import mean_squared_error
 
 ## just in case someone wants to implement them instead of using sklearn
 
+def rmse(X_test, Y_test, model):
+    Y_pred = model.predict(X_test)
+    return mean_squared_error(Y_test, Y_pred, squared=False)
+
+def plot_knn(ns, rmses):
+    plt.plot(ns, rmses, 'r*')
+    plt.xlabel('# of neighbors')
+    plt.ylabel('RMSE')
+    
+    plt.savefig('graphs/knn_choice_n.png', bbox_inches='tight')
+    plt.clf()
+
+def knn(xs, ys, n=10):
+    X_train, X_test, Y_train, Y_test = train_test_split(xs, ys, test_size= .1, random_state = 40)
+    num_cols = len(X_train.columns)
+    i = 5
+
+    best_model = KNeighborsRegressor(n_neighbors=i, algorithm='kd_tree').fit(X_train, Y_train)
+    best_rmse = rmse(X_test, Y_test, best_model)
+
+    ### Cross Validation
+    ns = [n]
+    rmses = [best_rmse]
+    ### You can change 5 to * 2 or * 3 here for a better result, but slower.
+    for n in range(i, int(np.sqrt(num_cols)) + 5):
+        model = KNeighborsRegressor(n_neighbors=n, algorithm='kd_tree').fit(X_train, Y_train)
+        temp = rmse(X_test, Y_test, model)
+        ns.append(n)
+        rmses.append(temp)
+        if temp < best_rmse:
+            best_model = model
+            best_rmse = temp
+    plot_knn(ns, rmses)
+
+    return best_model
+
+def gamma_model(xs, ys):
+    model = GammaRegressor().fit(xs, ys)
+    return model
+
+def linear_model(xs, ys, m):
+    model = make_pipeline(PolynomialFeatures(m), Ridge(normalize=True)).fit(xs, ys)
+    return model
+
+def random_forest(xs, ys):
+    model = RandomForestRegressor(criterion='mse').fit(xs, ys)
+    return model
+
+
+"""
 ## Modified KNeighborsRegressor so that it uses median rather than mean
+from sklearn.neighbors.regression import check_array, _get_weights
+
 class MedianKNeighborsRegressor(KNeighborsRegressor):
     def predict(self, X):
         X = check_array(X, accept_sparse='csr')
@@ -40,19 +94,4 @@ class MedianKNeighborsRegressor(KNeighborsRegressor):
 def medianknn(xs, ys, n):
     model = MedianKNeighborsRegressor(n_neighbors=n).fit(xs, ys)
     return model
-
-def knn(xs, ys, n):
-    model = KNeighborsRegressor(n_neighbors=n, algorithm='kd_tree').fit(xs, ys)
-    return model
-
-def gamma_model(xs, ys):
-    model = GammaRegressor().fit(xs, ys)
-    return model
-
-def linear_model(xs, ys, m):
-    model = make_pipeline(PolynomialFeatures(m), Ridge(normalize=True)).fit(xs, ys)
-    return model
-
-def random_forest(xs, ys):
-    model = RandomForestRegressor().fit(xs, ys)
-    return model
+"""
